+++ stm-servicos/src/main/java/io/github/franciscosviana/stmservicos/api/model/output/CredenciadoOutput.java
@@ -3,6 +3,7 @@ package io.github.franciscosviana.stmservicos.api.model.output;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -18,8 +19,8 @@ public class CredenciadoOutput {
     private String uf;
     private String tipo;
 
-    private double valorChamado;
-    private double valorKm;
+    private BigDecimal valorChamado;
+    private BigDecimal valorKm;
 
     private int quantidadeOSAtendidas;
